Fix series position not showing in UI
- Changed StorytellerSeriesApiModel to use @SerialName("position") instead of @SerialName("sequence") to match actual API response
- Added uuid and featured fields to StorytellerSeriesApiModel
- Updated toDomain() mapping to use position field and prefer uuid for series id
- Fixed type consistency in GetBooksBySeriesUseCase sorting (Double.MAX_VALUE instead of Int.MAX_VALUE)

+++ feature/books/domain/src/commonMain/kotlin/com/retro99/books/domain/usecase/GetBooksBySeriesUseCase.kt
@@ -21,7 +21,7 @@ class GetBooksBySeriesUseCase(
                     }.sortedWith(
                         compareBy(
                             { book ->
-                                book.series.find { it.name.equals(seriesName, ignoreCase = true) }?.position ?: Int.MAX_VALUE
+                                book.series.find { it.name.equals(seriesName, ignoreCase = true) }?.position ?: Double.MAX_VALUE
                             },
                             { it.title },
                         ),
